+++ open_manipulator/CHANGELOG.rst
@@ -2,6 +2,11 @@
 Changelog for package open_manipulator
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+4.0.7 (2025-07-17)
+------------------
+* Fixed joint offset functionality of joint trajectory command broadcaster
+* Contributors: Woojin Wie
+
 4.0.6 (2025-07-15)
 ------------------
 * Renamed omx to open_manipulator_x